# docs/swagger.yaml
definitions:
  common.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.PhotoCreateRequestSwaggo:
    properties:
      caption:
        example: example
        type: string
      photo_url:
        example: www.example.com
        type: string
      title:
        example: example
        type: string
    type: object
  model.PhotoUpdateRequestSwaggo:
    properties:
      caption:
        example: example
        type: string
      photo_url:
        example: www.example.com
        type: string
      title:
        example: example
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /photos:
    get:
      consumes:
      - application/json
      description: Get all photos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Get all photos
      tags:
      - Photos
    post:
      consumes:
      - application/json
      description: Create a new photo
      parameters:
      - description: Photo Create Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.PhotoCreateRequestSwaggo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Create a new photo
      tags:
      - Photos
  /photos/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete a photo
      parameters:
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Delete a photo
      tags:
      - Photos
    get:
      consumes:
      - application/json
      description: Get one photo
      parameters:
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Get one photo
      tags:
      - Photos
    put:
      consumes:
      - application/json
      description: Update a photo
      parameters:
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: string
      - description: Photo Update
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/model.PhotoUpdateRequestSwaggo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Response'
            type: array
      security:
      - Bearer: []
      summary: Update a photo
      tags:
      - Photos
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"